requirements.txt
google-generativeai>=0.3.0
lightgbm==4.6.0
beautifulsoup4==4.13.4
matplotlib>=3.5.0
pandas>=2.0.0
requests>=2.28.0
python-dotenv>=0.19.0
scikit-learn>=1.0.0
imbalanced-learn>=0.11.0
joblib>=1.2.0
flask>=2.0.0
flask-cors>=3.0.10
python-pptx==0.6.23
seaborn>=0.12.0
yagmail>=0.15.0